GPU使用

+++ server.py
@@ -42,7 +42,7 @@ strategy = fl.server.strategy.FedProx(
     fraction_fit=0.5,
     fraction_evaluate=1.0,
     evaluate_metrics_aggregation_fn=aggregate_metrics,
-    min_fit_clients=num_clients/2,         # 学習を開始するために最低10クライアントの参加を要求
+    min_fit_clients=int(num_clients / 2),         # 学習を開始するために最低10クライアントの参加を要求
     min_available_clients=num_clients,   # 少なくとも10クライアントが利用可能であることを要求
     proximal_mu=0.01,
 )
